+++ CHANGELOG.md
@@ -20,6 +20,7 @@ This project changes will be shown here.
 - Issue 467 - Leased line contract expiry notification
 - Issue 483 - Add Password level descriptions
 - Issue 480 - Current Projects shown on Sales Order page
+- Issue 489 - Contract Extraction
 
 ## [v2.1.4] - 2019-02-20
 ### Fixed